Update bat position in world map
Now the bat position is updated in the tile map. This will simplify the
ball's collision detection: the bat will be treated as another set of
tiles/blocks, not to mention it decouples the ball from the bat object.
This approach is not without it's drawbacks, So might reconsider
changing it in the near future.